src/third_party/blink/web_tests/external/wpt/svg/struct/scripted/autofocus-attribute-expected.txt

This is a testharness.js-based test.
FAIL <a> should support autofocus assert_equals: expected Element node <a href="#" autofocus="autofocus"></a> but got null
FAIL Renderable element with tabindex should support autofocus assert_equals: expected Element node <path d="M0,0h8v8z" tabindex="0" autofocus="autofocus"></... but got null
FAIL Never-rendered element with tabindex should not support autofocus promise_test: Unhandled rejection with value: object "TypeError: Cannot read property 'tagName' of null"
Harness: the test ran to completion.
